requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
jupyter>=1.0.0
scikit-learn==1.6.1
xgboost>=3.1.2
joblib>=1.2.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
Pillow>=9.0.0
